State: Move isRegeneratingJetpackPostByEmail to global selectors directory
