+++ client/state/jetpack/settings/selectors.js
@@ -4,18 +4,6 @@
 import { get } from 'lodash';
 
 /**
- * Returns true if we are currently making a request to regenerate the Post By Email address. False otherwise
- * Returns null if the status for the queried site is unknown.
- *
- * @param  {Object}  state       Global state tree
- * @param  {Number}  siteId      The ID of the site we're querying
- * @return {?Boolean}            Whether Post by Email address is currently being updated
- */
-export function isRegeneratingPostByEmail( state, siteId ) {
-	return get( state.jetpack.settings.requests, [ siteId, 'regeneratingPostByEmail' ], null );
-}
-
-/**
  * Returns the Jetpack settings on a certain site.
  * Returns null if the site is unknown, or settings have not been fetched yet.
  *
