+++ client/state/jetpack/settings/test/selectors.js
@@ -4,7 +4,6 @@
 import { expect } from 'chai';
 
 import {
-	isRegeneratingPostByEmail,
 	getJetpackSettings,
 	getJetpackSetting,
 	getJetpackSettingsSaveRequestStatus,
@@ -12,53 +11,9 @@ import {
 	getJetpackSettingsSaveError,
 } from '../selectors';
 
-import {
-	requests as REQUESTS_FIXTURE,
-	settings as SETTINGS_FIXTURE
-} from './fixture';
+import { settings as SETTINGS_FIXTURE } from './fixture';
 
 describe( 'selectors', () => {
-	describe( '#isRegeneratingPostByEmail', () => {
-		it( 'should return true if post by email is currently being regenerated', () => {
-			const stateIn = {
-					jetpack: {
-						settings: {
-							requests: REQUESTS_FIXTURE
-						}
-					}
-				},
-				siteId = 12345678;
-			const output = isRegeneratingPostByEmail( stateIn, siteId );
-			expect( output ).to.be.true;
-		} );
-
-		it( 'should return false if post by email is currently not being regenerated', () => {
-			const stateIn = {
-					jetpack: {
-						settings: {
-							requests: REQUESTS_FIXTURE
-						}
-					}
-				},
-				siteId = 87654321;
-			const output = isRegeneratingPostByEmail( stateIn, siteId );
-			expect( output ).to.be.false;
-		} );
-
-		it( 'should return null if that site is not known', () => {
-			const stateIn = {
-					jetpack: {
-						settings: {
-							requests: REQUESTS_FIXTURE
-						}
-					}
-				},
-				siteId = 88888888;
-			const output = isRegeneratingPostByEmail( stateIn, siteId );
-			expect( output ).to.be.null;
-		} );
-	} );
-
 	describe( '#getJetpackSettings', () => {
 		it( 'should return settings for all modules for a known site', () => {
 			const stateIn = {
