+++ client/state/selectors/index.js
@@ -48,6 +48,7 @@ export isJetpackSiteInDevelopmentMode from './is-jetpack-site-in-development-mod
 export isJetpackSiteInStagingMode from './is-jetpack-site-in-staging-mode';
 export isPrivateSite from './is-private-site';
 export isPublicizeEnabled from './is-publicize-enabled';
+export isRegeneratingJetpackPostByEmail from './is-regenerating-jetpack-post-by-email';
 export isRequestingBillingTransactions from './is-requesting-billing-transactions';
 export isRequestingJetpackConnectionStatus from './is-requesting-jetpack-connection-status';
 export isRequestingJetpackJumpstartStatus from './is-requesting-jetpack-jumpstart-status';
